unify definition of preimage as 32-byte slice
