+++ modules/fedimint-ln/src/contracts/incoming.rs
@@ -1,4 +1,4 @@
-use crate::contracts::{ContractId, IdentifyableContract};
+use crate::contracts::{ContractId, DecryptedPreimage, EncryptedPreimage, IdentifyableContract};
 use bitcoin_hashes::sha256::Hash as Sha256;
 use bitcoin_hashes::Hash as BitcoinHash;
 use bitcoin_hashes::{borrow_slice_impl, hash_newtype, hex_fmt_impl, index_impl, serde_impl};
@@ -72,17 +72,6 @@ pub struct FundedIncomingContract {
     pub out_point: OutPoint,
 }
 
-/// Possible outcomes of preimage decryption
-#[derive(Debug, Clone, Eq, PartialEq, Hash, Deserialize, Serialize, Encodable, Decodable)]
-pub enum DecryptedPreimage {
-    /// There aren't enough decryption shares yet
-    Pending,
-    /// The decrypted preimage was valid
-    Some(Preimage),
-    /// The decrypted preimage was invalid
-    Invalid,
-}
-
 hash_newtype!(
     OfferId,
     Sha256,
@@ -90,45 +79,9 @@ hash_newtype!(
     doc = "The hash of a LN incoming contract offer"
 );
 
-/// A preimage in the context of incoming contracts. In this context it is a public key chosen
-/// by the creator of the [`IncomingContractOffer`].
-#[derive(Debug, Clone, Eq, PartialEq, Hash, Deserialize, Serialize, Encodable, Decodable)]
-pub struct Preimage(pub secp256k1::XOnlyPublicKey);
-
-/// Threshold-encrypted [`Preimage`]
-#[derive(Debug, Clone, Eq, PartialEq, Hash, Deserialize, Serialize)]
-pub struct EncryptedPreimage(pub threshold_crypto::Ciphertext);
-
-/// Share to decrypt an [`EncryptedPreimage`]
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct PreimageDecryptionShare(pub threshold_crypto::DecryptionShare);
-
-impl EncryptedPreimage {
-    pub fn new(preimage: [u8; 32], key: &threshold_crypto::PublicKey) -> EncryptedPreimage {
-        EncryptedPreimage(key.encrypt(preimage))
-    }
-}
-
 impl IdentifyableContract for IncomingContract {
     fn contract_id(&self) -> ContractId {
         ContractId::from_hash(self.hash)
-    }
-}
-
-impl Encodable for EncryptedPreimage {
-    fn consensus_encode<W: std::io::Write>(&self, writer: W) -> Result<usize, Error> {
-        // TODO: get rid of bincode
-        let bytes = bincode::serialize(&self.0).expect("Serialization shouldn't fail");
-        bytes.consensus_encode(writer)
-    }
-}
-
-impl Decodable for EncryptedPreimage {
-    fn consensus_decode<D: std::io::Read>(d: D) -> Result<Self, DecodeError> {
-        let bytes = Vec::<u8>::consensus_decode(d)?;
-        Ok(EncryptedPreimage(
-            bincode::deserialize(&bytes).map_err(DecodeError::from_err)?,
-        ))
     }
 }
 
@@ -141,22 +94,5 @@ impl Encodable for OfferId {
 impl Decodable for OfferId {
     fn consensus_decode<D: std::io::Read>(d: D) -> Result<Self, DecodeError> {
         Ok(OfferId::from_inner(Decodable::consensus_decode(d)?))
-    }
-}
-
-impl Encodable for PreimageDecryptionShare {
-    fn consensus_encode<W: std::io::Write>(&self, writer: W) -> Result<usize, Error> {
-        // TODO: get rid of bincode
-        let bytes = bincode::serialize(&self.0).expect("Serialization shouldn't fail");
-        bytes.consensus_encode(writer)
-    }
-}
-
-impl Decodable for PreimageDecryptionShare {
-    fn consensus_decode<D: std::io::Read>(d: D) -> Result<Self, DecodeError> {
-        let bytes = Vec::<u8>::consensus_decode(d)?;
-        Ok(PreimageDecryptionShare(
-            bincode::deserialize(&bytes).map_err(DecodeError::from_err)?,
-        ))
     }
 }
